gw: update service interface

+++ src/main/kotlin/ru/dmitry4k/geomarkback/rest/dto/PostAreaMarkRequest.kt
@@ -1,6 +1,6 @@
 package ru.dmitry4k.geomarkback.rest.dto
 
 data class PostAreaMarkRequest(
-    val rate: Float,
+    val rate: Double,
     val points: List<GeoPointRequest>
 )
